Handle duplicated outputs using ExpressionEvaluator.
I also tried using output-to-output aliases to handle this. It didn't
work well because `getOutputAlias` couldn't distinguish the same `Val*`
at different indices. To allocate a tensor for the first output but not
the second, I would still end up with something like `outputs_map`.

This PR removes the ad-hoc `outputs_map` in favor of
`ExpressionEvaluator`.