adjust when should the persistent buffuers being projected to inputs (#1408)
Issue:
In normalization scheduler, persistent buffuers are only allowed to
project to inputs if the recalculation doesn't require reduction ops.
With this restriction, the persistent buffer in softmax is not being
projected to inputs and fp32 is required to store the buffter.

Fix:
(1) **Removal of the existing limitation**.
(2) To minimize redundant recalculations, the buffer is now **projected
onto the broadcasted tensor** following the initial reduction.
(3) Despite these changes facilitating the use of fp16 for persistent
buffers without necessitating extra reductions, a performance decline
has been observed in scenarios with smaller hidden sizes. Consequently,
the third alteration in this PR is the **deactivation of projection in
situations where the benefits of reducing buffer size do not outweigh
the drawbacks of recalculations**. Given the challenge in precisely
measuring these benefits and drawbacks prior to actual application, the
approach adopted in this PR is based purely on empirical data derived
from softmax experiments. It is implemented in a standalone funciton and
can be extended to handle other cases, e.g. RNG ops.

Performance:
(1) 50% increase for cases with hidden size large than 20K.
(2) 5% decrease for cases with hidden size in range of 10~12K. (hope can
be removed after revising the inner persistent heuristics)
(3) for outer softmax, up to 1.69x speedup is observed for reduction size of 1024