Mark aliases during scheduling.  (#1478)
This is one of the many steps to solve #1401.

Currently, we run alias analysis and mark aliases before segmentation.
This can't handle the output-to-output alias patterns we observed in the
nanogpt2 benchmark, because some schedulers fork outputs and therefore
break the aliasing chain observed before segmentation (and thus before
scheduling).
https://github.com/NVIDIA/Fuser/issues/1401#issuecomment-1840137527
shows an example for curious minds.

This PR changes the algorithm so it runs alias analysis both before
segmentation and during scheduling. The former is used to change layouts
to enable aliases, and the latter is used to detect and mark aliases in
each segment. This allows segmentation to split out meta-op-only regions
so the rest is easier to optimize.

This PR doesn't solve the aforementioned output-to-output alias
patterns. I'll need to add logic to segment out meta-op-only
output-to-output regions in later PRs.