Fix test_deterministic_random (#1523)
The python frontend test `test_deterministic_random` was originally
introduced in #546. It actually printed mismatches, but did not fail the
test in those cases. It turns out it has been failing but it went
unnoticed. This change unmasks that allclose check.

The test had assumed the offset should be advanced by 4 for every 4-byte
RNG op. This is how PyTorch handles offsets. However, we base our offset
on uint4, so our offset corresponds to 1/4 the pytorch offset. This
means we should only advance by one for each op in the test. Doing this
silences the warnings (which are now errors).

See https://github.com/NVIDIA/Fuser/blob/main/csrc/rng.cpp#L54-L64 for
more info on our offsets and their equivalents in PyTorch.