Allocate intermediate global buffers the same way we allocate outputs (#1469)
This fixes a subtle bug that was exposed by #973. In the test
`ExpandedBroadcastGlobalIntermediateTest.TheTest_CUDA`, which currently
passes, we are binding an unexpanded output tensor of size (2, 1, 2) to
the intermediate `T1_g` even though that tensor should have an expanded
size of `1 << 60 = 1152921504606846976`.
```
%kernel {
i10 = (nvfuser_index_t)(1152921504606846976);
T1_g[ iblockIdx.x3{2}, bS4{( (nvfuser_index_t)(1152921504606846976) )}, ithreadIdx.x5{2} ] = expand( T0_g[ iS0{2}, bS1{1}, iS2{2} ], {2, i10, 2} )
T2_g[ iblockIdx.x6{2}, bS7{( (nvfuser_index_t)(1152921504606846976) )}, ithreadIdx.x8{2} ]
   = Set( T1_g[ iblockIdx.x3{2}, bS4{( (nvfuser_index_t)(1152921504606846976) )}, ithreadIdx.x5{2} ], cache_op=Streaming )
```
The expression evaluator currently doesn't complain that we are binding
the value 1 to a constant int with value 1152921504606846976 since in
fact the expanded extent of that axis has a cast, i.e. it is the
non-immediate constant integer `(nvfuser_index_t)1152921504606846976`.
That is the actual `Val*` being bound, so the `ExpressionEvaluator` does
not complain that the values conflict, but once the cast is simplified
away in #973, we see the error.

This PR fixes this problem by allocating intermediate buffers in the
same way we allocate outputs, meaning we will also allocate expanded
tensors. `ExpressionEvaluator` is updated to check for expanded extents
when binding.

---------

Co-authored-by: Naoya Maruyama <naoyam@users.noreply.github.com>
Co-authored-by: Naoya Maruyama <nmaruyama@nvidia.com>