Add utilities for debugging MMA (#1518)
During development, I feel that it is super helpful to set one matrix as
identity, and set another matrix as arange, and print out the result. By
reading the result, I can easily tell the memory layout. I think this
debugging technique is so helpful that it worth committing it to our
codebase for future reference.