fix allocation of intermediate tensor in pipelines (#1244)
We provide a new function 
```
std::unordered_map<Val*, c10::IValue> allocatePipelineIntermediateBuffers(
    Pipeline* pipeline,
    DeviceIdxType my_device_index,
    std::vector<c10::IValue> global_inputs_IValues);
```
used to allocate intermediate tensors when executing Pipeline in
multidevice setting.

Before this patch, we would allocate a lot of intermediate tensors, more precisely, every output of each pipeline stage
(including global outputs), even if we don't run the stage. This results
in a lot of garbage allocation.

This function allocates only the required intermediate tensors, that
is, tensors that will be received from other devices through a network
communication and used as a subsequent stage's input. The allocation of
global or intermediate outputs is taken care of by the fusion executor.

we also provide a patch to fix the test after these changes