Fix logic error in PromoteReuseSyncModifier (#1381)
This fixes an issue that can appear when using smem for matmul epilogue.
It is possible to request promotion of prologue reuse and not wind up
re-using it currently. The reason is that we promote by inserting block
syncs in between lifetimes of smem allocations. The logic for this sync
insertion step currently assumes that there is an expression that
corresponds to each "position" as determined by `ScopeMap`. However,
`ScopeMap` actually increments the position at the end of each for loop,
in order to distinguish outer last reads from the last inner expression
position in the corresponding loop nest. We can wind up with an
inter-lifetime sync interval that exists entirely within these skipped
endfor positions, in which case the current code will not insert syncs,
meaning we do not re-use memory.

See the `SkippedSyncInterval` test, which fails on `main`, for an
example showing how parallelization of outer dimensions can shrink outer
lifetime intervals such that they get lost in the skipped endfor region.

This PR fixes this logic error by ensuring that any skipped position
corresponding to a last read (i.e. the start of an inter-lifetime
interval) is processed properly, translating the end of that interval to
the next non-skipped position in case the interval also ended within the
skipped range. It also checks that the ends of sync intervals are not
skipped.

---------

Co-authored-by: Gao, Xiang <qasdfgtyuiop@gmail.com>