Fix bug in codediff for pytest (#1544)
This should fix the `assert self.current_test is not None` errors we see
in CI sometimes.

Also, the default reference git ref was changed from `origin/main` to
`git merge-base origin/main HEAD`. This can help when a branch is not
up-to-date with upstream, since it prevents comparing against commits
that have not yet been merged into the branch.

I also added a warning when the number of kernels in a test changes. We
still compare the first N kernels where N is the minimum of the number
of kernels between the two runs, but now if that number differs we give
a warning. This helps alert us that we might not be comparing apples to
apples and we should look at that specific test in more detail manually.