Visit extent scalars in `SegmentCandidateFinder::resolveScalarsInGroup` (#840)
During segmentation, groups are defined as lists of `Expr`s. When
finalizing the segmentation, scalar inputs that might be needed are
sprinkled in using `resolveScalarsInGroup`. The method
`resolveScalarsInGroup` attempts to find all scalars needed in the group
and adds their defining exprs to the group; in this way scalars can be
recomputed for multiple groups if needed. When the `SegmentedGroup` is
finalized, all expression inputs in all the `Expr`s in the group are
added as inputs (deduplicated, removing constants, etc.).

Currently, `resolveScalarsInGroup` only processes the inputs of current
`Expr`s to look for input scalars using `input->isScalar()`. As seen in
#656, we also need to look at output extents in some cases. This PR
modifies that method to look also at intermediate IterDomain expressions
and their attributes, to more aggressively search for scalars.

Fixes #656.