Refactor topologicalSortStatements
For a given Value, all its Val dependencies require a valid definition
Expr. Initially, we create a Val without definition expression. This Val
is only valid for Expr. Then, we create an Expr that requires inputs,
outputs, and attribute Vals. After creating an Expr, we assign the Expr to
output Vals' definition. Now, output Val are valid for other Vals.