Handle dynamic ExpandOp during concretization (#1535)
This adds `ExpandOp` to the possible dynamic transforms we explicitly
handle during concretization, beside `ViewOp` and `Resize`.

`ExpandOp::checkConcretization` is not useful actually since the output
of the expand is not replaced until after that method is called. That
means we have no way to tell which axes are actually expanded, so we
cannot test whether the inputs should be `Broadcast` at that point.
Instead, this check is done at `analyzeExpands`, and the
`SymbolicExpand` test exercises it by passing in invalid sizes.

Since this adds a member to both `DynamicTransformInitialInfo` and
`DynamicTransformConcretizationInfo`, there are a number of mechanical
changes to cloning, comparison operator, hash(), etc. The important
changes are in `arith.cpp`, and in `dynamic_transform.cpp` in
`analyzeExpands` and `concretizeExpand`.

Fixes #1398. Fixes #1468.