Undo special handling for bothWays
Without splitk, we want to propagate like

  ... <- mma_result -> smem_epilogue

This terminates at smem_epilogue if we provide `{smem_epilogue}` as the
forward boundary.

Previously, I was modifying this to be `splitk_sum` when we had both
splitk_sum and smem_epilogue. However, that meant we were propagating

  ... <- mma_result -> smem_epilogue -> splitk_sum

terminating before splitk_sum. But this was transforming smem_epilogue,
which was undesired and led to a lack of swizzling in smem_epilogue.
Now, we simply revert to the previous case. In fact, we could probably
just propagate backwards and not use smem_epilogue at all since it seems
we never want to propagate forward here...

Note that if use_smem_epilogue == false, then smem_epilogue ==
mma_result are the same pointer so there will be no forward propagation
in that case.