Prevent replacing constant extents with non-const in concretization (#1575)
Fixes #1572

In that case, `concretizeReshape` is replacing some extents in the new
shape, such as `(nvfuser_index_t)16` with non-constant values like `i0`.
This leads to a downstream expression that originally was constant `0`
to be non-constant, so it cannot be picked up by `RemoveEmptyPass`. This
PR avoids concretizing constant extents with non-constants. This should
be safe since even if those tensors are segmentation edges, we will not
lose any scalars because the extent is constant so there are no free
scalars to lose.