Move `assoc_comm::FlattenedAssocCommOp` class declaration to its header file (#1601)
This PR moves the `assoc_comm::FlattenedAssocCommOp` class declaration
to the `expr_simplifier.h` header file. It also adds the appropriate
`handle` functions to `OptOutConstDispatch` and `OptOutDispatch`, so it
works with functions like `StmtSort::getStmts`. This PR is a
prerequisite for https://github.com/NVIDIA/Fuser/pull/1566, which aims to
serialize the entire IR.