IdModel: Add an initial loop graph (no promotion yet) (#1569)
This PR adds an initial loop graph. It should be effectively equivalent
to the current ComputeAt loop map except that non-leaf domains are also
mapped. That may not be strictly necessary but is useful for subsequent
loop promotion analysis, which will come after this PR.

I did not add any validation as we already validate the permissive graph
and this initial loop graph is pretty straightforward sub graph of the
permissive graph.