[DynamicType] Add named comparison ops returning dynamic type (#1619)
Fixes https://github.com/NVIDIA/Fuser/issues/1570

This PR adds a variant of comparison operators that returns
`DynamicType`. This new variant is not implemented as operator
overloading, instead, users need to explicitly call its function. The
reason for not overloading it as operators is because it makes more
sense to make comparison operators return bool, so that code like
```C++
PolymorphicValue pv = 2;
if (pv > 0) {
  blabla();
}
```
will seamlessly work.