add reduction comms (#1451)
Add a lowering path for reduction collective.
More precisely, before this patch: in a multi-device pipeline, the expr
in-between two stages was required to be a "set". With this patch, it
can also be a reduction. This expression will be lowered to MPI
collectives of the type `Reduce`, `Allreduce` or `ReduceScatter`.
We augment our test units to cover those scenarios.

Some restrictions:
- Only proper reduce-scatter without futher resharding are supported for
now (i.e. I/O of a reduce scatter must be sharded on the same device
mesh. If further resharding is desired, the user must manually insert a
set before/after the reduction to take care of the device mesh change)
- Only reducing one axis at a time is supported
- If a tensor is reduced in-between stage, it must be sharded on the
producer side (i.e., the reduction must occurs accross devices ;
otherwise the operation boils down to a local reduction + a
non-reduction resharding, and can be obtained by inserting a "set" after
the reduction op). Similarily, the sharding must be non-trivial, i.e.,
on an axis of size > 1, otherwise it is automatically replaced by a
"squeeze" and thus will not be supported.

Those restrictions will be relaxed later and handled automatically, but
it is more convenient and generic to handle those corner cases by
modifying the Fusion itself by adding "Set" exprs before or after the
reduction. This will be the object of a future PR.