IdModel: Fix inconsistent graph issue  (#1627)
Fixes #1636 

Previously when exprs are merged, their input and output groups are
updated to point to the new merged group. However, this is not enough as
there may be val groups that are used or defined by the pre-merged expr
groups but they may not be inputs or outputs of the exprs. See #1636 for
a concrete example.

Here, we check all mapping entries of `unique_definitions_` and
`unique_uses_` so all uses and definitions are correctly updated.