WIP: Build a map of broadcast resolutions for root domains
This is the first step of the loop promotion analysis. It may not be
very clear what this analysis is intended to do at this point, but in
short, the root resolution information is propagated to intermediate and
leaf ID groups in the IEL graph, which is then projected back to the
loop graph. I thought before diving into the full promotion analysis it
would make reviewing easier to split this part off as the first sub-PR
with unit tests.