Refactoring to allow more flexible testing (#1648)
Previously IdModel constructor builds all graphs at once. Now each graph
can be built individually. This refactoring is done in preparation for
loop promotion analysis to allow testing of each step of the analysis.
It would also be useful as sometimes not all graphs would be necessary
to build.

Stacked on #1627