Scheduling/lowering for serial grid reduction (#1456)
This change enables `ReductionOp`s to be lowered as serial reductions
(see #1405) if requested during scheduling.

1. At scheduling, a `ReductionOp` is modified by calling its
`requestSerialGridReduction()` method. The output tensor can be
scheduled before or after this method call, and should result in the op
having all its reduction axes parallelized as grid dimensions.
2. Early in lowering, we find `ReductionOp`s having
`serialGridReductionRequested() == true`, and we place syncs around
their outer loop. At this point, we also analyze that outer loop to
determine if there are any conflicting expressions, such as conflicting
grid reductions.
3. Later in lowering, during the indexing pass, we convert the
`ReductionOp` to a `GridReduction` that has its serial buffer set. The
serial buffer is a temporary `TensorIndex` indexed like a global memory
version of the reduction output tensor.

The generated kernel looks like this:
```c++
  // Allocate global tensor T4
  grid_sync::blockSerializeWait<false, false, true>(&T4[index_utils::maskedOffset<true, true, false>(blockIdx, gridDim)]);
  #pragma unroll
  for(nvfuser_index_t i13 = 0; i13 < 4LL; ++i13) {
    nvfuser_index_t i14;
    i14 = 8LL * i13;
    nvfuser_index_t i15;
    i15 = 2048LL * i13;
    nvfuser_index_t i16;
    i16 = i4 + i15;
    nvfuser_index_t i17;
    i17 = -i15;
    #pragma unroll
    for(nvfuser_index_t i18 = 0; i18 < 8LL; ++i18) {
      nvfuser_index_t i19;
      i19 = 256LL * (i18 + nvfuser_zero);
      nvfuser_index_t i20;
      i20 = i16 + i19;
      float T3[1LL];
      T3[0LL] = 0.000000000e+00f;
      // Allocate global tensor T5
      reduction::serialReductionStep(
        T3[0LL],
        T2[(i14 + i18)],
        0.000000000e+00f,
        T5[i20],
        [](float &a, float b) { a = a + b; },
        index_utils::maskedOffset<false, false, true>(blockIdx, gridDim) == 0,
        index_utils::maskedOffset<false, false, true>(blockIdx, gridDim) == index_utils::maskedSize<false, false, true>(gridDim) - 1,
        true,
        true);
      if ((b6 && (i5 < (i17 - i19)))) {
        T1[i20]
           = T3[0LL];
      }
    }
  }
  NVFUSER_UPDATE_MAGIC_ZERO;
  grid_sync::blockSerializeRelease<false, false, true>(&T4[index_utils::maskedOffset<true, true, false>(blockIdx, gridDim)]);
```
Notice that the index `i20` now matches between the output `T1` and the
intermediate `T5`. In another PR, I will attempt to extend our buffer
reuse machinery to recognize this as a chance to use `T1` in place of
`T5` (i.e. inner aliasing, in-place reduction).

Also notice that I have not yet hoisted the sync flags index, or the
`first_block` and `last_block` predicates.