Don't print entire CUDA source on compile error
When we arrive at a CUDA compilation error, we print the kernel along
with the entire preamble, which is thousands of lines. This blows away
the scrollback buffer in our terminal and is not typically very useful.
Kernel printing without preamble is still available via
`NVFUSER_DUMP=cuda_kernel` which occurs before this point so it is not
skipped on compile error. Unfortunately it is probably difficult to
automatically enable that option on compilation error.