[Hopper] MBarrier storage scope update (#1663)
The purpose of this change is to update the expected scope in which
smem placeholder for mbarrier objects are alive before the memory
can be reused. Life of such variable starts with its initialization, and
ends with its invalidation.

The core change is switching from last write to last read for mbarrier
invalidate helper function call.