Insert squeeze before trivial reductions at concretization (#1673)
This change inspects expressions at concretization to check for
reductions and Welford ops. When finding those, we look for unexpanded
broadcasts in the axes being reduced. If we find any, we re-run the
appropriate op with the concretized input, to ensure that the output is
squeezed exactly as it would be were the fusion not dynamic.

Fixes #1667

---------

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>