Add blank impl of isLowerableToCommunication.
isLowerableToCommunication is used in a few places to print error
messages or short-circuit loops. Those places appear to be places
that are intended to largely be used behind the distributed path.
It's easier to just define the API instead of trying to
conditionalize all the use sites and invent non-USE_DISTRIBUTED
behavior.