IdModel: Loop Promotion Analysis Step 1: Build a map of broadcast resolutions for root domains (#1650)
This is the first step of the [loop promotion
analysis](https://github.com/NVIDIA/Fuser/blob/IterDomainGraphs/csrc/id_model/id_model.cpp#L1032).
It may not be very clear what this analysis is intended to do at this
point, but in short, the root resolution information is propagated to
intermediate and leaf ID groups in the IEL graph, which is then
projected back to the loop graph. I thought before diving into the full
promotion analysis it would make reviewing easier to split this part off
as the first sub-PR with unit tests.

There's a detailed design doc for the overall loop promotion analysis.
DM me.