Target arch validation (#1784)
This introduces a minimum device version attribute to the kernel summary
and populates it by traversing the fusion early in lowering to find ops
or data types introduced after sm70. This version constraint is checked
in FusionExecutor::compileFusion and if the target arch doesn't satisfy
the constraint then an informative error message is given. Currently
BFloat16, cp.async and cp.async.bulk are checked, as are the various MMA
macros.

Fixes #1760

---------

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>