Hide symbols by default. (#1324)
... and require explicit marking to make them public. This gives us
explicit control over our ABI.

I recommend starting your review with `doc/dev/visibility.md`. This
describes what the purpose of the PR is and how we might update it in
the future.

The CMake changes are really the ones to scrutinize. There's lots of
code changes, but they are really just adding in `NVF_API` annotations
to a bunch of places, or some minor include/forward declaration changes.