Remove function test/utils.h/matmul (#1817)
This PR removes the `matmul` function in `test/utils.h`. In all unit
tests and benchmarks, I am replacing code like
```C++
auto c = matmul(a, b, layout);
```
with
```C++
a = canonicalizeInputToBMNK(a, layout, MmaOperand::A);
b = canonicalizeInputToBMNK(b, layout, MmaOperand::B);
auto c = fusedMultiplySum(a, b, {-1});
```
I believe the latter is better because it is more explicit about how the
fusion looks like, therefore more readable.

Besides, with this new coding style, it will be easier to define
supported prologue fusions for matmul. For example, if I want to do
`sin(A) x sin(B)`, I can just do:
```
a = canonicalizeInputToBMNK(a, layout, MmaOperand::A);
b = canonicalizeInputToBMNK(b, layout, MmaOperand::B);
auto c = fusedMultiplySum(sin(a), sin(b), {-1});
```
In the past, this was not possible (`auto c = matmul(sin(a), sin(b),
layout)` will not work, because of our current limitation that the
transposes must be acting on the fusion input tensor, because this
transpose will be converted to `ldmatrix.trans`)