Fix a segmentation bug with input forwarding.  (#1825)
This PR takes code from @naoyam's #1559 and restructures it. 

Before this PR, nvFuser resolves forwarded inputs by unconditionally
pulling in their dependencies into the group. This was problematic
(#1541) because a group may require a resize op to only appear as an
input. This PR fixes #1541 by checking the dependencies are indeed
mergeable before pulling them in.