Option to skip scheduling in FusionExecutorCache (#1844)
This PR is comprised of mainly two dependent parts:

### Skip scheduling

We add the ability in `FusionExecutorCache` to skip scheduling the
kernel with the heuristics.

This feature allows the user to use manual scheduling instead of
necessarily using the heuristic suggested by
the auto-scheduler.

This feature is enabled by passing a bool argument `auto_scheduling` to
`FusionExecutorCache`. This argument is optionnal and set to `true` by
default, thus the default behavior is not changed. We want to keep this
less permissive behavior as default to avoid uncaught errors.


### Use correct executor in `MultiDeviceExecutor`

Currently, we are using manual multi-device scheduling. Therefore the
above feature is necessary, since auto-scheduler is not compatible with
manual scheduling.

As a work around, we were using so far`FusionExecutor` in the
`MultiDeviceExecutor`. However, using `FusionExecutor`, is faulty here
(as pointed out by @wujingyue), and we need to use `FusionExecutorCache`
instead (which is anyway the main entry point for nvFuser runtime).
Thanks to the first feature of this pr, we are able to use
`FusionExecutorCache` inside `MultiDeviceExecutor`. Moreover, this patch
allows benefiting from the cache and other goodies that accelerate the
flow.

### TODOs
Later, we might wanna 1) modify all the existing scheduler to integrate
the multidevice aspect, and 2) get rid of `MultiDeviceExecutor` and
integrate it into `FusionExecutorCache` itself. In the meantime this
patch is relatively simple and allows us to go forward.