Disable castInputOutputToLowerPrecision by default.  (#1889)
At this very moment, `castInputOutputToLowerPrecision` seems to bring
more problems than benefits. So I think we'd better turn it off by
default. It has caused precision loss (#1742) and program crash (#1872).
It was added for performance, but the current framework benchmarks are
neutral to disabling it (the results are put in
`1889_benchmark_results.txt` in the nvFuser team's folder).

Properly fixing `castInputOutputToLowerPrecision` will take some time.
To properly fix #1742, it'll have to be made more conservative, i.e.,
adding a cast only when it preserves numeric equivalence. To properly
fix #1872, it'll have to stop adding a cast before a resizing op.

Because this PR disables `castInputOutputToLowerPrecision` by default,
it has to add option guards to some unit tests that exercise the
function.

Fixes #1742.

Fixes #1872.